Delete data from storage.

+++ Components/RewriteMe.Business/Services/StorageService.cs
@@ -99,13 +99,21 @@ namespace RewriteMe.Business.Services
             }
         }
 
-        public async Task DeleteFileItemDataAsync(Guid userId, Guid fileItemId)
+        public async Task DeleteFileItemFolderDataAsync(Guid userId, Guid fileItemId)
         {
             var container = await GetContainerClient(userId).ConfigureAwait(false);
-            var client = container.GetBlobClient(fileItemId.ToString());
-            await client.DeleteIfExistsAsync().ConfigureAwait(false);
+            var blobItems = container.GetBlobs().AsPages().SelectMany(x => x.Values).ToList();
 
-            _logger.Information($"File item container '{fileItemId}' was deleted.");
+            foreach (var blobItem in blobItems)
+            {
+                if (blobItem.Name.Contains(fileItemId.ToString(), StringComparison.InvariantCulture))
+                {
+                    var client = container.GetBlobClient(blobItem.Name);
+                    await client.DeleteIfExistsAsync().ConfigureAwait(false);
+                }
+            }
+
+            _logger.Information($"File item folder data '{fileItemId}' was deleted.");
         }
 
         public async Task DeleteFileItemSourceAsync(FileItem fileItem)
